Rename SubGroupSchedulingPolicy and make top-level
